Add an "inner hexagon"
I decided not to bother with automated integration tests for now, and instead to sketch out the interface to a model that can do the heavy-lifting work, which we can easily unit test.

I figured since @tooky and @andersjaensson don't know TypeScript so well, it might be more useful for me to put this structure in place.