Get the functional test working 🎉
I learned a lot about nock, and about GitHub authentication!

To generate the fixtures, I had to first generate a JWT so I could be
the app and use _that_ to generate an authentication token for an
installation, which then allowed me to query the repo!

I simplified the CHANGELOG.md so the fixture only contained one issue so
there was less to mock